requirements.txt
torch==2.7.0+cu128
torchvision==0.22.0+cu128
torchaudio==2.7.0+cu128
optimum==1.23.3
datasets==3.1.0
evaluate==0.4.3
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
tqdm==4.67.1
opencv-python==4.10.0.84
timm==0.9.12
decord==0.6.0
imageio==2.37.0
pillow==10.4.0
faiss-gpu==1.7.2
gradio==4.44.1
fastapi==0.115.6
uvicorn==0.32.1
openai==1.57.0
huggingface-hub==0.36.0
omegaconf==2.3.0
fire==0.7.0
PyYAML==6.0.1
requests==2.32.3
safetensors==0.4.5
regex==2024.11.6
tensorboard==2.19.0
matplotlib==3.9.3
seaborn==0.13.2
